[#4187] Move the project directory caching code to separate module
