+++ akvo/rest/views/project.py
@@ -20,8 +20,9 @@ from rest_framework.response import Response
 from rest_framework.status import HTTP_201_CREATED
 from geojson import Feature, Point, FeatureCollection
 
-from akvo.cache import get_cached_data, set_cached_data, cache_with_key
+from akvo.cache import get_cached_data, set_cached_data
 from akvo.codelists.store.default_codelists import SECTOR_CATEGORY
+from akvo.rest.cache import serialized_project
 from akvo.rest.serializers import (ProjectSerializer, ProjectExtraSerializer,
                                    ProjectExtraDeepSerializer,
                                    ProjectIatiExportSerializer,
@@ -37,7 +38,7 @@ from akvo.rest.models import TastyTokenAuthentication
 from akvo.rest.views.utils import (
     int_or_none, get_qs_elements_for_page
 )
-from akvo.rsr.models import Project, OrganisationCustomField, project_directory_cache_key
+from akvo.rsr.models import Project, OrganisationCustomField
 from akvo.rsr.filters import location_choices, get_m49_filter
 from akvo.rsr.views.my_rsr import user_viewable_projects
 from akvo.utils import codelist_choices
@@ -296,27 +297,6 @@ def project_directory_no_search(request):
     }
 
     return Response(response)
-
-
-# FIXME: Should we use a DB cache?
-@cache_with_key(project_directory_cache_key, timeout=None)
-def serialized_project(project_id):
-    project = Project.objects.only(
-        'id', 'title', 'subtitle',
-        'primary_location__id',
-        'primary_organisation__id',
-        'primary_organisation__name',
-        'primary_organisation__long_name'
-    ).select_related(
-        'primary_location',
-        'primary_organisation',
-    ).prefetch_related(
-        'locations',
-        'locations__country',
-        'recipient_countries',
-        'partners',
-    ).get(pk=project_id)
-    return ProjectDirectorySerializer(project).data
 
 
 @api_view(['GET'])
